LM cutting regions on import as well
